General Plugin decoupled Subviews still have to be decoupled
