+++ Library/RSBot.Core/BotCL.cs
@@ -1,16 +1,10 @@
 ﻿using RSBot.Core.Components;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
 
 namespace RSBot.Core
 {
     public class BotCL
     {
-        public static void Initialize(string profile, string silkroadDir)
+        public static void Initialize(string profile)
         {
             ProfileManager.SetSelectedProfile(profile);
             GlobalConfig.Load();
