--- Plugins/RSBot.Chat/Bootstrap.cs
@@ -1,42 +1,0 @@
-﻿//using System.Windows.Forms;
-//using RSBot.Core;
-//using RSBot.Core.Components;
-//using RSBot.Core.Plugins;
-
-//namespace RSBot.Chat;
-
-//public class Bootstrap : IPlugin
-//{
-//    /// <inheritdoc />
-//    public string InternalName => "RSBot.Chat";
-
-//    /// <inheritdoc />
-//    public string DisplayName => "Chat";
-
-//    /// <inheritdoc />
-//    public bool DisplayAsTab => true;
-
-//    /// <inheritdoc />
-//    public int Index => 98;
-
-//    /// <inheritdoc />
-//    public bool RequireIngame => true;
-
-//    /// <inheritdoc />
-//    public void Initialize() { }
-
-//    /// <inheritdoc />
-//    public Control View => Views.View.Instance;
-
-//    /// <inheritdoc />
-//    public void Translate()
-//    {
-//        LanguageManager.Translate(View, Kernel.Language);
-//    }
-
-//    /// <inheritdoc />a
-//    public void OnLoadCharacter()
-//    {
-//        // do nothing
-//    }
-//}
